requirements.txt
qrcode[pil]
pypng
pyzbar
pillow
opencv-python
numpy
pytelegrambotapi